requirements.txt
asgiref==3.7.2
Django==4.2.6
django-autoslug==1.9.9
django-environ==0.11.2
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
